Refactoring mapper implementations
